Move updating budget status to budget event handler
