+++ src/components/budget/handlers/index.ts
@@ -1,2 +1,3 @@
 export * from './create-project-default-budget.handler';
 export * from './sync-budget-records-to-funding-partners.handler';
+export * from './update-project-current-budget-status.handler';
